+++ test/Feature/CleanRegex/_prepared/inject/PatternTest.php
@@ -5,6 +5,7 @@ use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Test\Utils\Assertion\AssertsPattern;
 use TRegx\CleanRegex\Pattern;
+use TRegx\Exception\MalformedPatternException;
 
 class PatternTest extends TestCase
 {
@@ -328,5 +329,41 @@ class PatternTest extends TestCase
         // when, then
         $this->assertPatternTests($pattern, 'Foo:BAR');
         $this->assertPatternTests($pattern, 'Foo:bar');
+    }
+
+    /**
+     * @test
+     */
+    public function shouldThrowForGroupName_Placeholder()
+    {
+        // given
+        $pattern = Pattern::inject('(?<@>)', []);
+        // then
+        $this->expectException(MalformedPatternException::class);
+        $this->expectExceptionMessage('Subpattern name expected at offset 3');
+        // when
+        $pattern->test('Bar');
+    }
+
+    /**
+     * @test
+     */
+    public function shouldNotInjectIntoGroupName()
+    {
+        // given
+        $pattern = Pattern::inject('(?<@>)', []);
+        // when
+        $this->assertSame('/(?<@>)/', $pattern->delimited());
+    }
+
+    /**
+     * @test
+     */
+    public function shouldInjectIntoNegativeLookBehind()
+    {
+        // given
+        $pattern = Pattern::inject('(?<!@)', ['value']);
+        // when
+        $this->assertSame('/(?<!(?>value))/', $pattern->delimited());
     }
 }