+++ test/Supposition/provided.php
@@ -1,4 +1,7 @@
 <?php
+
+use Test\Utils\TestCase\DataProvider\NamedProvider;
+
 function provided(array $values): array
 {
     $provider = [];
@@ -10,10 +13,15 @@ function provided(array $values): array
 
 function named(array $values): array
 {
-    $provider = [];
-    foreach ($values as $value) {
-        $key = new \TRegx\CleanRegex\Internal\VisibleCharacters($value[0]);
-        $provider["$key"] = $value;
+    $provider = new NamedProvider();
+    foreach ($values as $key => $value) {
+        if (\is_array($value[0])) {
+            foreach ($value as $groupValue) {
+                $provider->addGroupEntry($key, $groupValue);
+            }
+        } else {
+            $provider->addEntry($value[0], $value);
+        }
     }
-    return $provider;
+    return $provider->toDataProvider();
 }